Simple calendar
This program first gets the current date using time() and localtime(). Then, it determines the number of days in the current month, including adjustments for leap years. Finally, it prints the calendar starting from the appropriate day of the week for the first day of the month.All this is mad possible through the use of <time.h> library.